Fixed testing errors
