+++ js/modules/magic-constants.js
@@ -1,0 +1,5 @@
+export const gameMagicConstants = {
+  FAST_ANSWER: 30,
+  MAX_CURRENT_ANSWERS: 10,
+  MAX_GAME_LEVEL: 10
+};
